pyproject.toml
[project]
name = "beam-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.3.0",
    "antlr4-python3-runtime>=4.13.2",
    "polars>=1.22.0",
    "rich>=13.9.4",
    "sympy>=1.13.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.48.3",
]
